rename: 파일 삭제 및 위치 변동 #116
