--- src/main/java/com/knu/sosuso/capstone/domain/video/service/TrendingService.java
@@ -1,131 +1,0 @@
-package com.knu.sosuso.capstone.domain.video.service;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.knu.sosuso.capstone.domain.video.dto.response.VideoSummaryResponse;
-import com.knu.sosuso.capstone.global.config.ApiConfig;
-import com.knu.sosuso.capstone.domain.detail.dto.DetailPageResponse;
-import com.knu.sosuso.capstone.global.exception.BusinessException;
-import com.knu.sosuso.capstone.global.exception.error.TrendingError;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class TrendingService {
-    private static final String YOUTUBE_VIDEOS_API_URL = "https://www.googleapis.com/youtube/v3/videos";
-
-    private final VideoProcessingService videoProcessingService;
-    private final ResponseMappingService responseMappingService;
-    private final ApiConfig config;
-    private final RestTemplate restTemplate;
-    private final ObjectMapper objectMapper;
-
-    public List<VideoSummaryResponse> getTrendingVideoWithComments(String token, String categoryType, int maxResults) {
-        log.info("인기급상승 영상 조회 시작: categoryType={}, maxResults={}", categoryType, maxResults);
-
-        try {
-            String categoryId = getCategoryId(categoryType);
-            List<String> videoIds = fetchTrendingVideoIds(categoryId, maxResults);
-
-            if (videoIds.isEmpty()) {
-                log.warn("인기급상승 영상이 없습니다: categoryType={}", categoryType);
-                return new ArrayList<>();
-            }
-
-            List<VideoSummaryResponse> results = new ArrayList<>();
-
-            for (String videoId : videoIds) {
-                try {
-                    log.debug("비디오 전체 처리 시작: apiVideoId={}", videoId);
-
-                    DetailPageResponse detailResponse = videoProcessingService
-                            .processVideoToSearchResult(token, videoId, true);
-
-                    if (detailResponse != null) {
-                        VideoSummaryResponse summaryResponse = responseMappingService
-                                .convertToVideoSummaryResponse(detailResponse);
-                        results.add(summaryResponse);
-                        log.debug("비디오 처리 완료: apiVideoId={}", videoId);
-                    } else {
-                        log.warn("비디오 처리 결과가 null: apiVideoId={}", videoId);
-                    }
-
-                } catch (Exception e) {
-                    log.error("개별 비디오 처리 실패: apiVideoId={}, error={}",
-                            videoId, e.getMessage(), e);
-                }
-            }
-
-            log.info("인기급상승 영상 조회 완료: 요청={}, 성공={}",
-                    videoIds.size(), results.size());
-            return results;
-
-        } catch (BusinessException e) {
-            throw e;
-
-        } catch (Exception e) {
-            log.error("인기급상승 영상 조회 실패: categoryType={}, error={}",
-                    categoryType, e.getMessage(), e);
-            throw new BusinessException(TrendingError.TRENDING_PROCESSING_ERROR);
-        }
-    }
-
-    private String getCategoryId(String categoryType) {
-        return switch (categoryType.toLowerCase()) {
-            case "latest" -> "0";
-            case "music" -> "10";
-            case "game" -> "20";
-            default -> throw new BusinessException(TrendingError.UNSUPPORTED_CATEGORY);
-        };
-    }
-
-    private List<String> fetchTrendingVideoIds(String categoryId, int maxResults) {
-        try {
-            String apiUrl = buildTrendingApiUrl(categoryId, maxResults);
-            log.debug("인기급상승 API 호출: {}", apiUrl);
-
-            String jsonResponse = restTemplate.getForObject(apiUrl, String.class);
-            JsonNode rootNode = objectMapper.readTree(jsonResponse);
-
-            List<String> videoIds = new ArrayList<>();
-            JsonNode itemsNode = rootNode.path("items");
-
-            if (itemsNode.isArray()) {
-                for (JsonNode item : itemsNode) {
-                    String videoId = item.path("id").asText();
-                    if (!videoId.isEmpty()) {
-                        videoIds.add(videoId);
-                    }
-                }
-            }
-
-            log.info("인기급상승 비디오 ID 조회 완료: 개수={}", videoIds.size());
-            return videoIds;
-
-        } catch (Exception e) {
-            log.error("인기급상승 비디오 ID 조회 실패: categoryId={}, error={}", categoryId, e.getMessage(), e);
-            throw new BusinessException(TrendingError.TRENDING_FETCH_ERROR);
-        }
-    }
-
-    private String buildTrendingApiUrl(String categoryId, int maxResults) {
-        UriComponentsBuilder builder = UriComponentsBuilder.fromUriString(YOUTUBE_VIDEOS_API_URL)
-                .queryParam("part", "id")
-                .queryParam("chart", "mostPopular")
-                .queryParam("regionCode", "KR")
-                .queryParam("maxResults", Math.min(maxResults, 30))  // YouTube API 제한
-                .queryParam("key", config.getKey());
-
-        builder.queryParam("videoCategoryId", categoryId);
-
-        return builder.build(false).toUriString();
-    }
-}
