+++ src/main/java/com/knu/sosuso/capstone/global/scheduler/VideoCleanupService.java
@@ -1,9 +1,10 @@
-package com.knu.sosuso.capstone.domain.video.service;
+package com.knu.sosuso.capstone.global.scheduler;
 
 import com.knu.sosuso.capstone.domain.video.entity.Video;
 import com.knu.sosuso.capstone.domain.video.repository.VideoRepository;
 import com.knu.sosuso.capstone.domain.scrap.repository.ScrapRepository;
 import com.knu.sosuso.capstone.domain.conmment.repository.CommentRepository;
+import com.knu.sosuso.capstone.domain.video.service.VideoProcessingService;
 import com.knu.sosuso.capstone.global.config.AppConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
